refactor(racing): refactor for feedback as Winner into Output, remove lines
