+++ src/main/java/racing/util/Splitter.java
@@ -1,4 +1,4 @@
-package racing.model;
+package racing.util;
 
 import java.util.Arrays;
 import java.util.List;
